[Fontations] Build bindgen targets, basic Rust test
Define meson bindgen steps that form a bindgen Rust library
that allows calling back to typical FontConfig functions for
constructing FcPatterns and adding them to an FcFontSet.

Add a basic Rust test that demonstrates it's possible to
call back to Fontconfig. Enable execution of this test
through:
    meson test -C build
as well as through:
    cargo test

Part of #434.