Deal with glob string properly
This is a regression of da1e3f35e61597fbf58c9c550ceba6d44abb1563

FcStrCopyFilename canonicalize a glob string though, the above
change expected to expand ~ character only to replace it with
a home directory.
However this also had a side-effect; adding a current directory
name if no home nor absolute path name in a glob string.

This isn't expected behavior here. So dropping it if there are,
otherwise use it as it is.

Fixes https://gitlab.freedesktop.org/fontconfig/fontconfig/-/issues/442