meson: don't force build of a shared library
Shared library linking has certain preconditions that may not always be
met, depending on usage. For example, static dependencies must be built
with PIC, which is unnecessary if we only require the static library.

We need to build both libraries only if `default_library` is not static.
If we are already building a static library, there is no need to link
the shared one. This also fixes the Windows build in this case, as
`c_shared_args` would otherwise force everything to be built twice.

This change is needed to fix the OSS-Fuzz build of FFmpeg.

Changelog: fixed
Fixes: 50aa6e3685223fb644b8800d5af77e1b6cda7345